test(driver-file): isolate child-abort lock test into its own binary
The coverage CI job runs `cargo test --all-features` under
cargo-llvm-cov instrumentation, which stretches the fork->exec window
long enough to surface a sibling-test fork-inheritance race in the
`lock` integration binary:

  - `open_after_drop_succeeds` holds a flock on its tempdir.
  - `open_after_child_abort_succeeds`, running in parallel in the same
    binary, calls `Command::spawn`. Between fork(2) and exec(2) the
    child inherits the parent's lock fd. Per flock(2), duplicated fds
    refer to the same lock; the OFD stays alive (and the lock stays
    held) until ALL such fds close.
  - The drop test then closes its fd and reopens, sees the lock still
    held by the not-yet-execed child, and gets EWOULDBLOCK.

Moving the child-spawning test into its own integration-test binary
removes the sibling-fd surface area entirely: cargo builds it as a
separate executable, and the binary contains only the one test, so the
spawn cannot fork-inherit any FileDriver fd from a concurrent sibling.

The driver-side fork-race fix (POSIX OFD locks, or closing the lock
fd around any spawn we do internally) is a separate concern -- the
driver does not spawn child processes today.
