+++ crates/tsoracle-driver-file/tests/lock.rs
@@ -13,10 +13,11 @@
 //! Single-writer enforcement via flock on the `LOCK` sentinel file.
 //!
 //! These tests pin the contract that `FileDriver::open_or_init` rejects a
-//! second concurrent open against the same directory, and that the OS-level
-//! lock is released both on graceful drop and on hard process abort.
+//! second concurrent open against the same directory, and that graceful
+//! Drop releases the lock. The kernel-cleanup-on-abort case lives in its
+//! own `lock_child_abort` binary — see that file for the fork-inheritance
+//! rationale.
 
-use std::process::Command;
 use tempfile::tempdir;
 use tsoracle_driver_file::{FileDriver, FileDriverError};
 
@@ -45,32 +46,4 @@ fn open_after_drop_succeeds() {
     drop(first);
     let _second = FileDriver::open_or_init(dir.path())
         .expect("second open after drop must succeed — drop releases the flock");
-}
-
-/// Child-process pattern: when `TSORACLE_LOCK_TEST_ABORT_DIR` is set, this
-/// test acts as the child — it opens the directory and `process::abort`s
-/// without unwinding. The kernel releases the flock on process death, so
-/// the parent must be able to reopen the same directory without seeing
-/// `AlreadyLocked`. Proves we don't depend on graceful Drop running.
-#[test]
-fn open_after_child_abort_succeeds() {
-    if let Ok(child_dir) = std::env::var("TSORACLE_LOCK_TEST_ABORT_DIR") {
-        let _driver = FileDriver::open_or_init(&child_dir).expect("child: open");
-        std::process::abort();
-    }
-
-    let dir = tempdir().unwrap();
-    let exe = std::env::current_exe().expect("current_exe");
-    let status = Command::new(&exe)
-        .args(["--exact", "--nocapture", "open_after_child_abort_succeeds"])
-        .env("TSORACLE_LOCK_TEST_ABORT_DIR", dir.path())
-        .status()
-        .expect("spawn child");
-    assert!(
-        !status.success(),
-        "child process should have aborted, got status {status:?}"
-    );
-
-    let _reopen = FileDriver::open_or_init(dir.path())
-        .expect("reopen after child abort must succeed — kernel released the flock");
 }
